+++ packages/toprf-secure-backup/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **Breaking**: Add `fetchMetadataAccessCreds` function option in `ToprfSecureBackup` class constructor.([#134]https://github.com/MetaMask/toprf-secure-backup/pull/134)
+  - This function should be used to inject metadata access token for calling metadata apis.
+
 ### Changed
 
 - Changed timestamp format from seconds to milliseconds in reset rate limit requests to minimize nonce collision. ([#133](https://github.com/MetaMask/toprf-secure-backup/pull/133))